move DockerContainerManager to separate folder
