+++ host_manager/docker_container_manager/container_configuration.py
